pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ap2"
version = "0.1.1"
description = "Agent Payments Protocol (Python)"
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.10"
keywords = ["payments", "a2a", "ap2"]
license = { text = "MIT" }
dependencies = [
  "pydantic>=2.6",
]
authors = [
  { name = "whill" }
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Libraries",
  "Topic :: Internet",
]

[project.urls]
Homepage = "https://github.com/whillhill/ap2"
Repository = "https://github.com/whillhill/ap2"
Issues = "https://github.com/whillhill/ap2/issues"
Documentation = "https://github.com/whillhill/ap2#readme"

[project.optional-dependencies]
samples = [
  "a2a-sdk",
  "google-adk",
  "google-genai",
  "httpx",
  "requests",
  "starlette",
  "uvicorn",
]
examples = [
  "a2a-sdk",
  "google-adk",
  "google-genai",
  "httpx",
  "requests",
  "starlette",
  "uvicorn",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["ap2*"]

[tool.setuptools]
license-files = ["LICENSE", "THIRD_PARTY_LICENSES/*"]
include-package-data = false
